+++ src/main/java/act/db/ebean/EbeanDao.java
@@ -1,4 +1,4 @@
-package act.db.ebean2;
+package act.db.ebean;
 
 /*-
  * #%L